refactor(parser): split FtlSpec tests into categories.
