+++ parser/test/src/io/github/fluent4s/parser/specs.scala
@@ -1,0 +1,12 @@
+package io.github.fluent4s.parser
+
+import org.scalatest._
+import flatspec._
+import matchers._
+
+abstract class UnitSpec
+    extends AnyFlatSpec
+    with should.Matchers
+    with OptionValues
+    with Inside
+    with Inspectors
